Add user handlers to Mirage config.js
